requirements.txt
# FastAPI and related packages
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.11.3
email_validator==2.2.0
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# Text and Speech Processing
gtts==2.5.4
langdetect==1.0.9

# AI and ML
pinecone==3.0.2
sentence-transformers==2.2.2
groq==0.23.1
langchain-groq==0.3.2
langchain-core==0.3.56
langsmith==0.3.37
transformers==4.30.2
torch==2.0.1
torchvision==0.15.2
huggingface-hub==0.16.4

# Additional Dependencies
requests==2.32.3
requests-toolbelt==1.0.0
python-dateutil==2.9.0
numpy==1.24.3
scikit-learn==1.6.1
scipy==1.15.2
Pillow==11.2.1
tqdm==4.67.1
regex==2024.11.6
pyyaml==6.0.2
safetensors==0.5.3
filelock==3.18.0
jinja2==3.1.6
sympy==1.14.0
networkx==3.4.2
fsspec==2025.3.2
distro==1.9.0

# Security and Utilities
bcrypt==4.3.0
cryptography==44.0.2
certifi==2025.4.26
charset-normalizer==3.4.1
dnspython==2.7.0
ecdsa==0.19.1
orjson==3.10.16
rsa==4.9.1
zstandard==0.23.0
